Refactor BiodataForm and update templates to Handlebars
