+++ frontend/src/components/common/ErrorMessage.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const ErrorMessage = ({ error }) => {
+  if (!error) return null;
+  return <p className="text-red-500 text-sm mt-1">{error.message}</p>;
+};
+
+export default ErrorMessage;
